+++ src/main/java/org/wise/portal/dao/project/impl/HibernateTagDao.java
@@ -54,10 +54,10 @@ public class HibernateTagDao extends AbstractHibernateDao<Tag> implements TagDao
 	 * @see org.wise.portal.dao.project.TagDao#getTagByName(java.lang.String)
 	 */
 	@SuppressWarnings("unchecked")
-	public Tag getTagByName(String name){
+	public Tag getTagByName(String name) {
 		List<Tag> tags = (List<Tag>) this.getHibernateTemplate().find("select tag from TagImpl tag where tag.name='" + name + "'");
 		
-		if(tags.size() == 0){
+		if (tags.size() == 0) {
 			return null;
 		} else {
 			return tags.get(0);
@@ -69,14 +69,14 @@ public class HibernateTagDao extends AbstractHibernateDao<Tag> implements TagDao
 	 * @param tagId
 	 */
 	@SuppressWarnings("unchecked")
-	public void removeIfOrphaned(Long tagId){
+	public void removeIfOrphaned(Integer tagId) {
 		List<Tag> projects = (List<Tag>) this.getHibernateTemplate().find("select project from ProjectImpl project inner join project.tags tag where tag.id=" + tagId);
 		
-		if(projects.size() == 0){
+		if (projects.size() == 0) {
 			try{
 				Tag tag = this.getById(tagId);
 				this.delete(tag);
-			} catch (ObjectNotFoundException e){
+			} catch (ObjectNotFoundException e) {
 				e.printStackTrace();
 			}
 		}
