+++ src/main/java/org/wise/portal/presentation/web/controllers/TaggerController.java
@@ -51,59 +51,59 @@ public class TaggerController {
 			HttpServletResponse response) throws Exception {
 		String command = request.getParameter("command");
 		
-		if(command.equals("createTag")){
+		if (command.equals("createTag")) {
 			String projectId = request.getParameter("projectId");
 			Project project = this.projectService.getById(Long.parseLong(projectId));
 			String tag = request.getParameter("tag");
 			
-			if(projectId == null || tag == null || projectId.equals("") || tag.equals("")){
+			if (projectId == null || tag == null || projectId.equals("") || tag.equals("")) {
 				response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Invalid parameters provided, cannot complete the request.");
-			} else if(this.projectService.projectContainsTag(project, tag)){
+			} else if (this.projectService.projectContainsTag(project, tag)) {
 				response.getWriter().write("duplicate");
-			} else if(!this.projectService.isAuthorizedToCreateTag(ControllerUtil.getSignedInUser(), tag)){
+			} else if (!this.projectService.isAuthorizedToCreateTag(ControllerUtil.getSignedInUser(), tag)) {
 				response.getWriter().write("not-authorized");
 			} else {
-				Long tagId = this.projectService.addTagToProject(tag, Long.parseLong(projectId));
+				Integer tagId = this.projectService.addTagToProject(tag, Long.parseLong(projectId));
 				response.getWriter().write(String.valueOf(tagId));
 			}
-		} else if(command.equals("removeTag")){
+		} else if (command.equals("removeTag")) {
 			String projectId = request.getParameter("projectId");
 			String tagId = request.getParameter("tagId");
 			
-			if(projectId == null || tagId == null || projectId.equals("") || tagId.equals("")){
+			if (projectId == null || tagId == null || projectId.equals("") || tagId.equals("")) {
 				response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Invalid parameters provided, cannot complete the request.");
 			} else {
-				this.projectService.removeTagFromProject(Long.parseLong(tagId), Long.parseLong(projectId));
+				this.projectService.removeTagFromProject(Integer.parseInt(tagId), Long.parseLong(projectId));
 				response.getWriter().write("success");
 			}
-		} else if(command.equals("updateTag")){
+		} else if (command.equals("updateTag")) {
 			String projectId = request.getParameter("projectId");
 			Project project = this.projectService.getById(Long.parseLong(projectId));
 			String tagId = request.getParameter("tagId");
 			String name = request.getParameter("name");
 			
-			if(projectId == null || tagId == null || name == null || projectId.equals("") || tagId.equals("") || name.equals("")){
+			if (projectId == null || tagId == null || name == null || projectId.equals("") || tagId.equals("") || name.equals("")) {
 				response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Invalid parameters provided, cannot complete the request.");
-			} else if(this.projectService.projectContainsTag(project, name)){
+			} else if (this.projectService.projectContainsTag(project, name)) {
 				response.getWriter().write("duplicate");
-			} else if(!this.projectService.isAuthorizedToCreateTag(ControllerUtil.getSignedInUser(), name)){
+			} else if (!this.projectService.isAuthorizedToCreateTag(ControllerUtil.getSignedInUser(), name)) {
 				response.getWriter().write("not-authorized");
 			} else {
-				Long id = this.projectService.updateTag(Long.parseLong(tagId), Long.parseLong(projectId), name);
+				Integer id = this.projectService.updateTag(Integer.parseInt(tagId), Long.parseLong(projectId), name);
 				response.getWriter().write(String.valueOf(id));
 			}
-		} else if(command.equals("retrieveProjectTags")){
+		} else if (command.equals("retrieveProjectTags")) {
 			String projectId = request.getParameter("projectId");
 			
-			if(projectId != null){
+			if (projectId != null) {
 				Project p = this.projectService.getById(Long.parseLong(projectId));
-				if(p != null){
+				if (p != null) {
 					String tagString = "";
-					for(Tag t : p.getTags()){
+					for(Tag t : p.getTags()) {
 						tagString += t.getId() + "~" + t.getName() + ",";
 					}
 					
-					/* remove trailing comma */
+					// remove trailing comma
 					tagString = tagString.substring(0,tagString.length() - 1);
 					
 					response.getWriter().write(tagString);
